Remove deprecated HistoryBuilder methods, move HistoryBuilder in History (#166)
